+++ NEWS.md
@@ -1,5 +1,8 @@
 # prenomfrance (development version)
 
+# prenomfrance 0.0.3
+* Ajout onglet Popularité avec nuage des prénoms dynamique et tableaux top10.
+* Ajout onglet Reseau: à partir d'un prenom choisi, affichage du reseau des prenoms proches selon la distance de Jaro Winkler.
 
 # prenomfrance 0.0.2
 * Ajout graphique de concentration des prénoms et bouton switch entre les deux graphiques.